feat: move stage color mapping to commonConstants
