+++ frontend/src/community/crm/constants/commonConstants.ts
